refact: separated the doc controllers into admin manager
M  src/main/java/com/tuvarna/phd/controller/DoctoralCenterAdminController.java
A  src/main/java/com/tuvarna/phd/service/DoctoralCenterAdminService.java
A  src/main/java/com/tuvarna/phd/service/DoctoralCenterAdminServiceImpl.java
M  src/main/java/com/tuvarna/phd/service/DoctoralCenterService.java
M  src/main/java/com/tuvarna/phd/service/DoctoralCenterServiceImpl.java
