+++ src/main/java/com/tuvarna/phd/service/DoctoralCenterService.java
@@ -2,9 +2,7 @@ package com.tuvarna.phd.service;
 
 import com.tuvarna.phd.dto.CandidateDTO;
 import com.tuvarna.phd.dto.CandidateStatusDTO;
-import com.tuvarna.phd.dto.RoleDTO;
 import com.tuvarna.phd.dto.UnauthorizedUsersDTO;
-import com.tuvarna.phd.dto.UserDTO;
 import com.tuvarna.phd.entity.UnauthorizedUsers;
 import java.io.IOException;
 import java.util.List;
@@ -17,11 +15,5 @@ public sealed interface DoctoralCenterService permits DoctoralCenterServiceImpl 
 
   void setUnauthorizedUserGroup(List<UnauthorizedUsersDTO> usersDTO, String group);
 
-  void changeUnauthorizedUserIsAllowed(String oid, Boolean isAllowed);
-
   List<UnauthorizedUsers> getUnauthorizedUsers();
-
-  List<UserDTO> getAuthorizedUsers();
-
-  void deleteAuthorizedUser(String oid, RoleDTO role);
 }
